2FEB:Shanoor - Created New Controllers for FreeLanceOnService and UserServiceDetails.
