+++ src/main/java/com/src/dao/NewServiceDAOImpl.java
@@ -14,12 +14,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.src.constant.AppConfig;
 import com.src.constant.NewServiceConstant;
-import com.src.constant.UserConstant;
-import com.src.entity.FreeLanceOnServiceEntity;
 import com.src.entity.NewServiceEntity;
 import com.src.entity.NewServiceHistoryEntity;
 import com.src.entity.NewServicePackageEntity;
-import com.src.entity.UserServiceDetailsEntity;
 import com.src.exception.RestCustomException;
 
 /**
@@ -117,125 +114,4 @@ public class NewServiceDAOImpl extends AbstractDAOManager implements NewServiceD
 				applicationConfigProperties.getProperty(AppConfig.GETALLUSERS_NOUSERSFOUND_ERRORMSG));
 	}
 
-	/**
-	 * To save the User Service Details.
-	 */
-	@Transactional
-	public UserServiceDetailsEntity saveUserServiceDetails(UserServiceDetailsEntity userServiceDetailsEntity) {
-		LOGGER.info(NewServiceConstant.INSIDE_SAVE_USERSERVICE_DETAILS);
-		int userServiceDetailsID = (Integer) sessionFactory.getCurrentSession().save(userServiceDetailsEntity);
-		LOGGER.debug(NewServiceConstant.CONFIRM_SAVE_USERSERVICE_DETAILS + userServiceDetailsID);
-		if (userServiceDetailsID > 0) {
-			return userServiceDetailsEntity;
-		}
-		throw new RestCustomException(HttpStatus.BAD_REQUEST,
-				applicationConfigProperties.getProperty(AppConfig.NEWSERVICE_UNABLE_TO_SAVE_ERRORMSG));
-	}
-
-	/**
-	 * To save the freelance on Service.
-	 */
-	@Transactional
-	public FreeLanceOnServiceEntity saveFreeLanceOnService(FreeLanceOnServiceEntity freeLanceOnServiceEntity) {
-		LOGGER.info(NewServiceConstant.INSIDE_SAVE_FREELANCE_ONSERVICE);
-		int userServiceDetailsID = (Integer) sessionFactory.getCurrentSession().save(freeLanceOnServiceEntity);
-		LOGGER.debug(NewServiceConstant.CONFIRM_SAVE_FREELANCE_ONSERVICE + userServiceDetailsID);
-		if (userServiceDetailsID > 0) {
-			return freeLanceOnServiceEntity;
-		}
-		throw new RestCustomException(HttpStatus.BAD_REQUEST,
-				applicationConfigProperties.getProperty(AppConfig.NEWSERVICE_UNABLE_TO_SAVE_ERRORMSG));
-	}
-
-	/**
-	 * To save or Update the New Service Details.
-	 */
-	@Transactional
-	public void saveOrUpdateUserSVCDetails(UserServiceDetailsEntity userServiceDetailsEntity) {
-		try {
-			LOGGER.info(NewServiceConstant.INSIDE_SAVEORUPDATE_USERSERVICEDETAILS);
-			sessionFactory.getCurrentSession().saveOrUpdate(userServiceDetailsEntity);
-			LOGGER.info(NewServiceConstant.CONFIRMED_SAVEORUPDATE_USERSERVICEDETAILS
-					+ userServiceDetailsEntity.getServiceId());
-		} catch (RestCustomException e) {
-			throw new RestCustomException(HttpStatus.BAD_REQUEST, applicationConfigProperties
-					.getProperty(AppConfig.SAVEORUPDATESERVICEDETAILS_UNABLETOUPDATE_ERRORMSG));
-		}
-	}
-
-	/**
-	 * To save or Update the New Service Details.
-	 */
-	@Transactional
-	public void saveOrUpdateFreeLanceOnService(FreeLanceOnServiceEntity freeLanceOnServiceEntity) {
-		try {
-			LOGGER.info(NewServiceConstant.INSIDE_SAVEORUPDATE_FREELANCE_ON_SERVICE);
-			sessionFactory.getCurrentSession().saveOrUpdate(freeLanceOnServiceEntity);
-			LOGGER.info(NewServiceConstant.CONFIRMED_SAVEORUPDATE_FREELANCE_ON_SERVICE
-					+ freeLanceOnServiceEntity.getJobId());
-		} catch (RestCustomException e) {
-			throw new RestCustomException(HttpStatus.BAD_REQUEST, applicationConfigProperties
-					.getProperty(AppConfig.SAVEORUPDATESERVICEDETAILS_UNABLETOUPDATE_ERRORMSG));
-		}
-	}
-
-	/**
-	 * Get the User Service Details by UserId.
-	 * 
-	 * @param userId
-	 */
-	@Transactional
-	public UserServiceDetailsEntity getUserServiceDetailsByUserId(int userId) {
-		LOGGER.info(UserConstant.USER_SERVICE_DAO_GETUSERSERVICEBYUSERID);
-		UserServiceDetailsEntity userServiceDetailsEntity = null;
-		Criteria criteria = this.sessionFactory.getCurrentSession().createCriteria(UserServiceDetailsEntity.class);
-		criteria.add(Restrictions.eq(UserConstant.USERID, userId));
-		userServiceDetailsEntity = (UserServiceDetailsEntity) criteria.uniqueResult();
-		if (userServiceDetailsEntity != null) {
-			return userServiceDetailsEntity;
-		}
-		throw new RestCustomException(HttpStatus.NO_CONTENT,
-				applicationConfigProperties.getProperty(AppConfig.GETUSERBYUSERID_USERNOTFOUND_ERRORMSG)
-						+ " for user Id : " + userId);
-	}
-
-	/**
-	 * Get the User Service Details by serviceId.
-	 * 
-	 * @param serviceId
-	 */
-	@Transactional
-	public UserServiceDetailsEntity getUserServiceDetailsByServiceId(int serviceId) {
-		LOGGER.info(UserConstant.USER_SERVICE_DAO_GETUSERSERVICEBYSERVICEID);
-		UserServiceDetailsEntity userServiceDetailsEntity = null;
-		Criteria criteria = this.sessionFactory.getCurrentSession().createCriteria(UserServiceDetailsEntity.class);
-		criteria.add(Restrictions.eq(UserConstant.SERVICEID, serviceId));
-		userServiceDetailsEntity = (UserServiceDetailsEntity) criteria.uniqueResult();
-		if (userServiceDetailsEntity != null) {
-			return userServiceDetailsEntity;
-		}
-		throw new RestCustomException(HttpStatus.NO_CONTENT,
-				applicationConfigProperties.getProperty(AppConfig.GETUSERBYUSERID_USERNOTFOUND_ERRORMSG)
-						+ " for Service Id : " + serviceId);
-	}
-
-	/**
-	 * To Get All User Service Details.
-	 */
-	@SuppressWarnings("unchecked")
-	@Transactional
-	public ArrayList<UserServiceDetailsEntity> getAllUserServiceDetails() {
-		LOGGER.info(NewServiceConstant.NEW_SERVICE_DAO_GETALLUSERSERVICEDETAILS);
-		List<UserServiceDetailsEntity> newUserServiceEntity = null;
-		Criteria criteria = this.sessionFactory.getCurrentSession().createCriteria(UserServiceDetailsEntity.class);
-		criteria.add(Restrictions.eq(NewServiceConstant.USER_SERVICE_DETAILS_ISACTIVE,true));
-		newUserServiceEntity = criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY).list();
-		int size = newUserServiceEntity != null ? newUserServiceEntity.size() : 0;
-		LOGGER.debug(NewServiceConstant.NEW_SERVICE_DAO_INSIDE_GETALLUSERSERVICEDETAILS + size);
-		if (size > 0) {
-			return (ArrayList<UserServiceDetailsEntity>) newUserServiceEntity;
-		}
-		throw new RestCustomException(HttpStatus.NO_CONTENT,
-				applicationConfigProperties.getProperty(AppConfig.GETALLUSERS_NOUSERSFOUND_ERRORMSG));
-	}
 }
