+++ src/main/java/com/src/dao/UserServiceDetailsDAO.java
@@ -1,0 +1,53 @@
+package com.src.dao;
+
+import java.util.ArrayList;
+
+import com.src.entity.UserServiceDetailsEntity;
+
+/**
+ * The <code> UserServiceDetailsDAO </code> interface defines a data access
+ * object for retrieving <code>User Service Details</code>.
+ * 
+ * @author Shanoor
+ *
+ */
+public interface UserServiceDetailsDAO {
+
+	/**
+	 * To save User Service Details.
+	 * 
+	 * @param UserServiceDetailsEntity
+	 * @return
+	 */
+	public UserServiceDetailsEntity saveUserServiceDetails(UserServiceDetailsEntity userServiceDetailsEntity);
+
+	/**
+	 * To edit and update the user Service details.
+	 * 
+	 * @param userServiceDetailsEntity
+	 * @return
+	 */
+	public void saveOrUpdateUserSVCDetails(UserServiceDetailsEntity userServiceDetailsEntity);
+
+	/**
+	 * Get the User Service Details by UserId.
+	 * 
+	 * @param userId
+	 */
+	public UserServiceDetailsEntity getUserServiceDetailsByUserId(int userId);
+
+	/**
+	 * Get the User Service Details by serviceId.
+	 * 
+	 * @param serviceId
+	 */
+	public UserServiceDetailsEntity getUserServiceDetailsByServiceId(int serviceId);
+
+	/**
+	 * To Get the all User Service Details.
+	 * 
+	 * @return Get All User Service Details.
+	 */
+	public ArrayList<UserServiceDetailsEntity> getAllUserServiceDetails();
+
+}
