+++ src/main/java/com/src/restcontroller/FreeLanceOnServiceController.java
@@ -1,0 +1,51 @@
+package com.src.restcontroller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import com.src.entity.FreeLanceOnServiceEntity;
+
+/**
+ * The <code> FreeLanceOnServiceController </code> class defines managed beans
+ * which provides back-end functionality on the
+ * <code>Free Lance on Service.</code>
+ * 
+ * @author Shanoor
+ *
+ */
+@Controller
+public class FreeLanceOnServiceController extends AbstractRestManager {
+
+	/**
+	 * Method is to save Free Lance On Service Details.
+	 * 
+	 * @param newServiceHistoryObject
+	 * @return
+	 */
+	@RequestMapping(value = "/saveFreeLanceOnService/", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<FreeLanceOnServiceEntity> saveFreeLanceOnService(
+			@RequestBody FreeLanceOnServiceEntity freeLanceOnServiceEntityObject) {
+		FreeLanceOnServiceEntity freeLanceOnServiceEntity = freeLanceOnServiceSVC
+				.saveFreeLanceOnService(freeLanceOnServiceEntityObject);
+		return new ResponseEntity<FreeLanceOnServiceEntity>(freeLanceOnServiceEntity, HttpStatus.OK);
+	}
+
+	/**
+	 * Method is to save or Update the Free Lance On Service.
+	 * 
+	 * @param saveOrUpdateFreeLanceOnService
+	 * @return
+	 */
+	@RequestMapping(value = "/saveOrUpdateFreeLanceOnService/", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<FreeLanceOnServiceEntity> saveOrUpdateFreeLanceOnService(
+			@RequestBody FreeLanceOnServiceEntity freeLanceOnServiceEntityObject) {
+		FreeLanceOnServiceEntity freeLanceOnServiceEntity = freeLanceOnServiceSVC
+				.saveOrUpdateFreeLanceOnService(freeLanceOnServiceEntityObject);
+		return new ResponseEntity<FreeLanceOnServiceEntity>(freeLanceOnServiceEntity, HttpStatus.OK);
+	}
+}
