+++ src/main/java/com/src/service/NewSVCService.java
@@ -2,11 +2,9 @@ package com.src.service;
 
 import java.util.ArrayList;
 
-import com.src.entity.FreeLanceOnServiceEntity;
 import com.src.entity.NewServiceEntity;
 import com.src.entity.NewServiceHistoryEntity;
 import com.src.entity.NewServicePackageEntity;
-import com.src.entity.UserServiceDetailsEntity;
 
 /**
  * The <code> NewSVCService </code> interface defines a business service for
@@ -57,56 +55,4 @@ public interface NewSVCService {
 	 */
 	public ArrayList<NewServiceEntity> getAllServiceDetails();
 
-	/**
-	 * To save User Service Details.
-	 * 
-	 * @param UserServiceDetailsEntity
-	 * @return
-	 */
-	public UserServiceDetailsEntity saveUserServiceDetails(UserServiceDetailsEntity userServiceDetailsEntity);
-
-	/**
-	 * To save Free Lance on Service.
-	 * 
-	 * @param FreeLanceOnServiceEntity
-	 * @return
-	 */
-	public FreeLanceOnServiceEntity saveFreeLanceOnService(FreeLanceOnServiceEntity freeLanceOnServiceEntity);
-
-	/**
-	 * To edit and update the user Service details.
-	 * 
-	 * @param userServiceDetailsEntity
-	 * @return
-	 */
-	public UserServiceDetailsEntity saveOrUpdateUserSVCDetails(UserServiceDetailsEntity userServiceDetailsEntity);
-
-	/**
-	 * To edit and Update the Free Lance on Service.
-	 * 
-	 * @param freeLanceOnServiceEntity
-	 * @return
-	 */
-	public FreeLanceOnServiceEntity saveOrUpdateFreeLanceOnService(FreeLanceOnServiceEntity freeLanceOnServiceEntity);
-
-	/**
-	 * Get the User Service Details by UserId.
-	 * 
-	 * @param userId
-	 */
-	public UserServiceDetailsEntity getUserServiceDetailsByUserId(int userId);
-
-	/**
-	 * Get the User Service Details by serviceId.
-	 * 
-	 * @param serviceId
-	 */
-	public UserServiceDetailsEntity getUserServiceDetailsByServiceId(int serviceId);
-
-	/**
-	 * To Get the all User Service Details.
-	 * 
-	 * @return Get All User Service Details.
-	 */
-	public ArrayList<UserServiceDetailsEntity> getAllUserServiceDetails();
 }
