+++ src/main/java/com/src/service/NewSVCServiceImpl.java
@@ -6,12 +6,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.src.constant.NewServiceConstant;
-import com.src.entity.FreeLanceOnServiceEntity;
 import com.src.entity.NewServiceEntity;
 import com.src.entity.NewServiceHistoryEntity;
 import com.src.entity.NewServicePackageEntity;
-import com.src.entity.UserServiceDetailsEntity;
-import com.src.entity.UserServiceEventHistoryEntity;
 import com.src.utils.CommonUtilites;
 
 /**
@@ -94,100 +91,5 @@ public class NewSVCServiceImpl extends AbstractServiceManager implements NewSVCS
 	@Override
 	public ArrayList<NewServiceEntity> getAllServiceDetails() {
 		return newServiceRestDAO.getAllServiceDetails();
-	}
-
-	/**
-	 * To Save User Service Details.
-	 */
-	@Override
-	public UserServiceDetailsEntity saveUserServiceDetails(UserServiceDetailsEntity userServiceDetailsEntity) {
-		userServiceDetailsEntity.setCreatedOn(CommonUtilites.getCurrentDateInNewFormat());
-		userServiceDetailsEntity.setActive(Boolean.TRUE);
-
-		NewServiceEntity newServiceEntity = new NewServiceEntity();
-		newServiceEntity.setOurserviceId(userServiceDetailsEntity.getOurserviceId());
-		userServiceDetailsEntity.setNewService(newServiceEntity);
-
-		UserServiceEventHistoryEntity userServiceEventHistory = userServiceDetailsEntity.getUserServiceEventHistory();
-		userServiceEventHistory.setUpdatedOn(CommonUtilites.getCurrentDateInNewFormat());
-		userServiceEventHistory.setUserServiceDetailsEntity(userServiceDetailsEntity);
-		userServiceDetailsEntity.setUserServiceEventHistory(userServiceEventHistory);
-		return newServiceRestDAO.saveUserServiceDetails(userServiceDetailsEntity);
-	}
-
-	/**
-	 * To save Free Lance On Service Details.
-	 */
-	@Override
-	public FreeLanceOnServiceEntity saveFreeLanceOnService(FreeLanceOnServiceEntity freeLanceOnServiceEntity) {
-		freeLanceOnServiceEntity.setUpdatedOn(CommonUtilites.getCurrentDateInNewFormat());
-		freeLanceOnServiceEntity.setJobAccepted(Boolean.FALSE);
-		freeLanceOnServiceEntity.setOutStandingAmtPaid(Boolean.FALSE);
-		freeLanceOnServiceEntity.setJobCancel(Boolean.FALSE);
-		freeLanceOnServiceEntity.setJobActive(Boolean.FALSE);
-		freeLanceOnServiceEntity.setJobCreatedOn(CommonUtilites.getCurrentDateInNewFormat());
-		UserServiceDetailsEntity userServiceDetailsEntity = new UserServiceDetailsEntity();
-		userServiceDetailsEntity.setServiceId(freeLanceOnServiceEntity.getServiceId());
-		freeLanceOnServiceEntity.setUserServiceDetails(userServiceDetailsEntity);
-		return newServiceRestDAO.saveFreeLanceOnService(freeLanceOnServiceEntity);
-	}
-
-	/**
-	 * To save or Update User Service Details.
-	 */
-	@Override
-	public UserServiceDetailsEntity saveOrUpdateUserSVCDetails(UserServiceDetailsEntity userServiceDetailsEntity) {
-		NewServiceEntity newServiceEntity = new NewServiceEntity();
-		newServiceEntity.setOurserviceId(userServiceDetailsEntity.getOurserviceId());
-		userServiceDetailsEntity.setNewService(newServiceEntity);
-
-		UserServiceEventHistoryEntity userServiceEventHistory = userServiceDetailsEntity.getUserServiceEventHistory();
-		userServiceEventHistory.setUpdatedOn(CommonUtilites.getCurrentDateInNewFormat());
-		userServiceEventHistory.setUserServiceDetailsEntity(userServiceDetailsEntity);
-		userServiceDetailsEntity.setUserServiceEventHistory(userServiceEventHistory);
-		newServiceRestDAO.saveOrUpdateUserSVCDetails(userServiceDetailsEntity);
-		return userServiceDetailsEntity;
-	}
-
-	/**
-	 * To save or Update FreeLance on Service Details.
-	 */
-	@Override
-	public FreeLanceOnServiceEntity saveOrUpdateFreeLanceOnService(FreeLanceOnServiceEntity freeLanceOnServiceEntity) {
-		freeLanceOnServiceEntity.setUpdatedOn(CommonUtilites.getCurrentDateInNewFormat());
-
-		UserServiceDetailsEntity userServiceDetails = new UserServiceDetailsEntity();
-		userServiceDetails.setServiceId(freeLanceOnServiceEntity.getServiceId());
-		freeLanceOnServiceEntity.setUserServiceDetails(userServiceDetails);
-		newServiceRestDAO.saveOrUpdateFreeLanceOnService(freeLanceOnServiceEntity);
-		return freeLanceOnServiceEntity;
-	}
-
-	/**
-	 * Get the User Service Details by UserId.
-	 * 
-	 * @param userId
-	 */
-	@Override
-	public UserServiceDetailsEntity getUserServiceDetailsByUserId(int userId) {
-		return newServiceRestDAO.getUserServiceDetailsByUserId(userId);
-	}
-
-	/**
-	 * Get the User Service Details by serviceId.
-	 * 
-	 * @param serviceId
-	 */
-	@Override
-	public UserServiceDetailsEntity getUserServiceDetailsByServiceId(int serviceId) {
-		return newServiceRestDAO.getUserServiceDetailsByServiceId(serviceId);
-	}
-
-	/**
-	 * To Get All User Service Details.
-	 */
-	@Override
-	public ArrayList<UserServiceDetailsEntity> getAllUserServiceDetails() {
-		return newServiceRestDAO.getAllUserServiceDetails();
 	}
 }
